Rename dataquest modue4-7 to dataquest module4-7
1. For Facebook, Instagram, and Minecraft: Pocket Edition, create lists that contain the name of the app (track_name), the rating count (rating_count_tot), and the user rating (user_rating). Remember, the first item in a list is at index 0.
a. For Facebook, assign the list to a variable named fb_rating_data.
b. For Instagram, assign the list to a variable named insta_rating_data.
c. For Minecraft: Pocket Edition, assign the list to a variable named minecraft_rating_data.

2. Use the data in fb_rating_data, insta_rating_data, and minecraft_rating_data to find the average user rating for these three apps.
a. Add the three user ratings (user_rating) together and save the sum as total_rating.
b. Divide the total (total_rating) by 3 to get the average rating. Save it as average_rating.