Create dataquest module4-9
Create a new list of lists that stores the lists (rows) for the applications that are games, Clash of Clans, Fruit Ninja Classic, and Minecraft: Pocket Edition. Assign the output to a variable named games_apps.

Use the variable names for each list generated by the display code in the code editor.